Move classes into org.skraylabs.poker.outcome package.
Create a new Java package for:
* Outcome
* OutcomeCalculator (formerly ProbabilityCalculator)
* OutcomeChecker
* WinLossCounter

These classes are all related and I didn’t really
like them being in org.skraylabs.poker package.

Also moves associated test classes.
