+++ src/test/java/com/skraylabs/poker/ApplicationFormatOutputTest.java
@@ -3,6 +3,9 @@ package com.skraylabs.poker;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertThat;
 
+import com.skraylabs.poker.outcome.Outcome;
+import com.skraylabs.poker.outcome.OutcomeCalculator;
+
 import org.junit.Before;
 import org.junit.Test;
