+++ src/test/java/com/skraylabs/poker/outcome/FlushTest.java
@@ -1,4 +1,4 @@
-package com.skraylabs.poker;
+package com.skraylabs.poker.outcome;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -9,30 +9,31 @@ import com.skraylabs.poker.model.GameState;
 import com.skraylabs.poker.model.GameStateFactory;
 import com.skraylabs.poker.model.GameStateFormatException;
 import com.skraylabs.poker.model.PocketFormatException;
+import com.skraylabs.poker.outcome.OutcomeCalculator;
 
 import org.junit.Test;
 
-public class TwoPairTest {
+public class FlushTest {
 
   @Test
   public void givenBustedHandReturnsZeroProbability() throws BoardFormatException,
       PocketFormatException, GameStateFormatException, CardFormatException {
-    GameState game = GameStateFactory.createGameStateFromString("As Kc Qh 9h\n 5s 2c");
+    GameState game = GameStateFactory.createGameStateFromString("As Ac Ah Kh\n Ts Js");
     OutcomeCalculator calculator = new OutcomeCalculator(game);
 
-    double probability = calculator.twoPairForPlayer(0);
+    double probability = calculator.flushForPlayer(0);
 
     assertThat(probability, equalTo(0.0));
   }
 
   @Test
-  public void givenAPairAndOneChanceReturnsCorrectProbability() throws BoardFormatException,
+  public void givenFlushDrawAndTwoChancesReturnsCorrectProbability() throws BoardFormatException,
       PocketFormatException, GameStateFormatException, CardFormatException {
-    GameState game = GameStateFactory.createGameStateFromString("As Kc Qh 9h\n 5s 5c");
+    GameState game = GameStateFactory.createGameStateFromString("As 2s Kh\n 4s 5s");
     OutcomeCalculator calculator = new OutcomeCalculator(game);
 
-    double probability = calculator.twoPairForPlayer(0);
+    double probability = calculator.flushForPlayer(0);
 
-    assertThat(probability, equalTo(12.0 / 46.0));
+    assertThat(probability, equalTo(378.0 / 1081.0));
   }
 }
