+++ src/test/java/com/skraylabs/poker/outcome/StraightTest.java
@@ -1,4 +1,4 @@
-package com.skraylabs.poker;
+package com.skraylabs.poker.outcome;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -9,30 +9,31 @@ import com.skraylabs.poker.model.GameState;
 import com.skraylabs.poker.model.GameStateFactory;
 import com.skraylabs.poker.model.GameStateFormatException;
 import com.skraylabs.poker.model.PocketFormatException;
+import com.skraylabs.poker.outcome.OutcomeCalculator;
 
 import org.junit.Test;
 
-public class FullHouseTest {
+public class StraightTest {
 
   @Test
-  public void givenBustedHandReturnsZeroProbability() throws CardFormatException,
+  public void givenABustedStraightReturnsZeroProbability() throws CardFormatException,
       BoardFormatException, PocketFormatException, GameStateFormatException {
-    GameState game = GameStateFactory.createGameStateFromString("Ah 2h 3h\n Kc Qc");
+    GameState game = GameStateFactory.createGameStateFromString("2s 3s 4s 9c\n 9h 9d");
     OutcomeCalculator calculator = new OutcomeCalculator(game);
 
-    double probability = calculator.fullHouseForPlayer(0);
+    double probability = calculator.straightForPlayer(0);
 
     assertThat(probability, equalTo(0.0));
   }
 
   @Test
-  public void givenTwoPairWithTwoChancesReturnsCorrectProbability() throws CardFormatException,
+  public void givenAStraightDrawWithOneChanceReturnsCorrectProbability() throws CardFormatException,
       BoardFormatException, PocketFormatException, GameStateFormatException {
-    GameState game = GameStateFactory.createGameStateFromString("Ah Kh Qh\n Kc Qc");
+    GameState game = GameStateFactory.createGameStateFromString("2s 3s 4s 9c\n 9h 5d");
     OutcomeCalculator calculator = new OutcomeCalculator(game);
 
-    double probability = calculator.fullHouseForPlayer(0);
+    double probability = calculator.straightForPlayer(0);
 
-    assertThat(probability, equalTo(181.0 / 1081.0));
+    assertThat(probability, equalTo(8.0 / 46.0));
   }
 }
