+++ src/test/java/com/skraylabs/poker/outcome/TwoPairTest.java
@@ -1,4 +1,4 @@
-package com.skraylabs.poker;
+package com.skraylabs.poker.outcome;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -9,30 +9,31 @@ import com.skraylabs.poker.model.GameState;
 import com.skraylabs.poker.model.GameStateFactory;
 import com.skraylabs.poker.model.GameStateFormatException;
 import com.skraylabs.poker.model.PocketFormatException;
+import com.skraylabs.poker.outcome.OutcomeCalculator;
 
 import org.junit.Test;
 
-public class FlushTest {
+public class TwoPairTest {
 
   @Test
   public void givenBustedHandReturnsZeroProbability() throws BoardFormatException,
       PocketFormatException, GameStateFormatException, CardFormatException {
-    GameState game = GameStateFactory.createGameStateFromString("As Ac Ah Kh\n Ts Js");
+    GameState game = GameStateFactory.createGameStateFromString("As Kc Qh 9h\n 5s 2c");
     OutcomeCalculator calculator = new OutcomeCalculator(game);
 
-    double probability = calculator.flushForPlayer(0);
+    double probability = calculator.twoPairForPlayer(0);
 
     assertThat(probability, equalTo(0.0));
   }
 
   @Test
-  public void givenFlushDrawAndTwoChancesReturnsCorrectProbability() throws BoardFormatException,
+  public void givenAPairAndOneChanceReturnsCorrectProbability() throws BoardFormatException,
       PocketFormatException, GameStateFormatException, CardFormatException {
-    GameState game = GameStateFactory.createGameStateFromString("As 2s Kh\n 4s 5s");
+    GameState game = GameStateFactory.createGameStateFromString("As Kc Qh 9h\n 5s 5c");
     OutcomeCalculator calculator = new OutcomeCalculator(game);
 
-    double probability = calculator.flushForPlayer(0);
+    double probability = calculator.twoPairForPlayer(0);
 
-    assertThat(probability, equalTo(378.0 / 1081.0));
+    assertThat(probability, equalTo(12.0 / 46.0));
   }
 }
